+++ src/aui-form-builder/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-1803](https://issues.liferay.com/browse/AUI-1803) Should be able to add and resize columns on main mode
 * [AUI-1785](https://issues.liferay.com/browse/AUI-1785) Merge list and choice fields into a single field type
 * [AUI-1800](https://issues.liferay.com/browse/AUI-1800) Add option on field toolbar to add a nested field
 * [AUI-1791](https://issues.liferay.com/browse/AUI-1791) Changing a field's content should update its row height